refactor: entry point into app and index
